pyproject.toml
[tool.poetry]
name = "clipboard-watcher"
version = "0.0.1"
description = "Monitors the access of other processes to the clipboard contents."
authors = ["Gonçalo Valério <gon@ovalerio.net>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
python-xlib = "^0.31"
dbus-python = "^1.2.18"
psutil = "^5.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.12b0", allow-prereleases = true}

[tool.poetry.scripts]
watcher = "clipboard_watcher.app:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
